refactor: CreateTeamServiceに変更、古い定義に基づくハードコーディングを削除
